+++ service/wechatopen/datacube.getweanalysisappiddailyvisittrend.go
@@ -9,15 +9,15 @@ import (
 
 type DataCubeGetWeAnAlySisAppidDailyVisitTrendResponse struct {
 	List []struct {
-		RefDate         string  `json:"ref_date"`
-		SessionCnt      int     `json:"session_cnt"`
-		VisitPv         int     `json:"visit_pv"`
-		VisitUv         int     `json:"visit_uv"`
-		VisitUvNew      int     `json:"visit_uv_new"`
-		StayTimeUv      float64 `json:"stay_time_uv"`
-		StayTimeSession float64 `json:"stay_time_session"`
-		VisitDepth      float64 `json:"visit_depth"`
-	} `json:"list"`
+		RefDate         string  `json:"ref_date"`          // 时间
+		SessionCnt      int64   `json:"session_cnt"`       // 打开次数（自然周内汇总）
+		VisitPv         int64   `json:"visit_pv"`          // 访问次数（自然周内汇总）
+		VisitUv         int64   `json:"visit_uv"`          // 访问人数（自然周内去重）
+		VisitUvNew      int64   `json:"visit_uv_new"`      // 新用户数（自然周内去重）
+		StayTimeUv      float64 `json:"stay_time_uv"`      // 人均停留时长 (浮点型，单位：秒)
+		StayTimeSession float64 `json:"stay_time_session"` // 次均停留时长 (浮点型，单位：秒)
+		VisitDepth      float64 `json:"visit_depth"`       // 平均访问深度 (浮点型)
+	} `json:"list"` // 数据列表
 }
 
 type DataCubeGetWeAnAlySisAppidDailyVisitTrendResult struct {